feat: consolidate the 3 SOQL E2E tests into 1 big test
